[REF] Move lap formatting to cli target
